separate domHnadling from index / added sass loader
