+++ src/assets/css/style.scss
@@ -1,7 +1,7 @@
 body {
   width:100vw;
   height: 90vh;
-  background-color: #eee;
+  background-color: rgb(0, 0, 0);
   display: flex;
   justify-content: center;
 }
