--- src/dom/userPage.js
@@ -1,6 +1,0 @@
-import start from '../index';
-
-
-
-
-
