+++ webpack.config.js
@@ -29,10 +29,14 @@ module.exports = {
         use: "raw-loader"
       },
       {
-        test: /\.css$/,
+        test: /\.s[ac]ss$/i,
         use: [
+          // Creates `style` nodes from JS strings
           'style-loader',
+          // Translates CSS into CommonJS
           'css-loader',
+          // Compiles Sass to CSS
+          'sass-loader',
         ],
       },
       {
